requirements.txt
pyrogram
tgcrypto
pymongo
python-dotenv
binpan
